+++ docs/.vitepress/config.js
@@ -1,7 +1,7 @@
 import { getSideBar } from "vitepress-plugin-autobar"
 export default {
-  title: "helpdoc", // 设置网站标题
-  description: "a online helpdoc",
+  title: "zhi-log", // 设置网站标题
+  description: "a simple logger for Node and Browser",
   base: "/", // 设置站点根路径
   outDir: "./.vitepress/dist", // 设置输出目录
   repo: "https://github.com/terwer/zhi-env", // 添加 git 链接
@@ -12,7 +12,7 @@ export default {
     // 添加导航栏
     nav: [
       {
-        text: "helpdoc",
+        text: "zhi-log",
         link: "/",
         target: "_self",
       },